fixing warning
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@5983 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ testing/testing_zgetrf_batched.cpp
@@ -220,7 +220,7 @@ int main( int argc, char** argv)
                     for(int k=0;k<min_mn; k++){
                         if(ipiv[i*min_mn+k] < 1 || ipiv[i*min_mn+k] > M )
                         {
-                                printf("error for matrix %d ipiv @ %d = %d\n",i,k,ipiv[i*min_mn+k]);
+                                printf("error for matrix %d ipiv @ %d = %d\n",i,k,(int)ipiv[i*min_mn+k]);
                                 stop=1;
                         }
                     }
